refactor: optimize parameterSecurityChecker
