+++ server/home/home-service/src/main/java/io/holoinsight/server/home/biz/service/TenantInitService.java
@@ -3,15 +3,13 @@
  */
 package io.holoinsight.server.home.biz.service;
 
-import io.holoinsight.server.apm.common.model.specification.sw.Tag;
+import io.holoinsight.server.common.dao.entity.CloudMonitorRange;
+import io.holoinsight.server.common.dao.entity.dto.IntegrationGeneratedDTO;
 import io.holoinsight.server.common.dao.entity.dto.MetricInfoDTO;
 import io.holoinsight.server.common.dao.entity.dto.TenantOpsStorage;
-import io.holoinsight.server.home.biz.plugin.config.MetaLabel;
 import io.holoinsight.server.common.scope.MonitorScope;
 import io.holoinsight.server.common.scope.MonitorUser;
-import io.holoinsight.server.common.dao.entity.CloudMonitorRange;
-import io.holoinsight.server.home.dal.model.dto.CustomPluginDTO;
-import io.holoinsight.server.common.dao.entity.dto.IntegrationGeneratedDTO;
+import io.holoinsight.server.home.biz.plugin.config.MetaLabel;
 import io.holoinsight.server.query.grpc.QueryProto.QueryFilter;
 
 import java.util.List;
@@ -101,21 +99,13 @@ public interface TenantInitService {
   CloudMonitorRange getCollectMonitorRange(String table, String tenant, String workspace,
       List<String> strings, MetaLabel metaLabel);
 
-  Boolean checkCookie(String tenant, String workspace, String environment);
-
-  Boolean checkTraceTags(String tenant, String workspace, List<Tag> tags);
-
-  Boolean checkTraceParams(String tenant, String workspace, Map<String, String> paramsMap);
 
   List<IntegrationGeneratedDTO> getExtraGeneratedLists();
-
-  Boolean checkIntegrationWorkspace(String workspace);
-
-  Boolean checkCustomPluginLogConfParams(String tenant, String workspace,
-      CustomPluginDTO customPluginDTO);
 
   List<String> getAggCompletenessTags();
 
   List<String> getAggDefaultGroupByTags();
+
+  Boolean checkIntegrationWorkspace(String workspace);
 
 }
