+++ server/home/home-web/src/main/java/io/holoinsight/server/home/web/security/ParameterSecurityService.java
@@ -3,8 +3,10 @@
  */
 package io.holoinsight.server.home.web.security;
 
+import io.holoinsight.server.apm.common.model.specification.sw.Tag;
 import io.holoinsight.server.common.model.DataQueryRequest;
 import io.holoinsight.server.common.scope.MonitorUser;
+import io.holoinsight.server.home.dal.model.dto.CustomPluginDTO;
 
 import java.util.List;
 import java.util.Map;
@@ -40,4 +42,20 @@ public interface ParameterSecurityService {
   boolean checkTenant(String target, String tenant);
 
   boolean checkSourceId(Long sourceId, String sourceType, String tenant, String workspace);
+
+  Boolean checkAgentLogPath(String logpath);
+
+  Boolean checkAgentLogPathPrefix(String logpath);
+
+
+  Boolean checkCookie(String tenant, String workspace, String environment);
+
+  Boolean checkTraceTags(String tenant, String workspace, List<Tag> tags);
+
+  Boolean checkTraceParams(String tenant, String workspace, Map<String, String> paramsMap);
+
+
+
+  Boolean checkCustomPluginLogConfParams(String tenant, String workspace,
+      CustomPluginDTO customPluginDTO);
 }
